Main App/requirements.txt
Flask
requests
kafka-python
pandas
numpy
cassandra-driver
flask-wtf
cqlengine
schedule